Solve RadioGroup issues
